refactor: remove empty files, move content to utils.ts
