--- server/src/utils/arraysOperationUtiil.ts
@@ -1,3 +1,0 @@
-export const shuffleArray = <T = unknown>(array: Array<T>) => {
-  return array.sort(() => 0.5 - Math.random());
-};
