--- server/src/utils/getBaseLogUtil.ts
@@ -1,3 +1,0 @@
-export const getBaseLog = (x: number, y: number) => {
-  return Math.log(y) / Math.log(x);
-};
