--- server/src/utils/getLastNItemsMapUtil.ts
@@ -1,3 +1,0 @@
-export const getLastNItemsFromMap = (map: Map<unknown, unknown>, n = 0) => {
-  return map ? Array.from(map.entries()).slice(-n) : [];
-};
